+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mwubneh <mwubneh@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 14:22:24 by mwubneh           #+#    #+#             */
-/*   Updated: 2023/01/17 13:34:18 by mwubneh          ###   ########lyon.fr   */
+/*   Updated: 2023/01/17 14:25:44 by mwubneh          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ size_t		ft_putnbr(int64_t nbr);
 int			ft_abs(int nbr);
 size_t		ft_strlen(const char *str);
 int			ft_print_point(va_list arg);
-int			ft_print_str(const char *str);
+size_t		ft_print_str(const char *str);
 int			ft_print_char(char c);
 int			ft_printf(const char *str, ...);
 size_t		ft_putnbr_hexa(uint64_t nbr, char *base);
